% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/as_list.R
\name{as.list.xml_nodeset}
\alias{as.list.xml_nodeset}
\title{Coerce xml nodes to a list.}
\usage{
\method{as.list}{xml_nodeset}(x, ns = character(), ...)
}
\arguments{
\item{x}{A document, node, or node set.}

\item{ns}{Optionally, a named vector giving prefix-url pairs, as produced
by \code{\link{xml_ns}}. If provided, all names will be explicitly
qualified with the ns prefix, i.e. if the element \code{bar} is defined
in namespace \code{foo}, it will be called \code{foo:bar}. (And
similarly for atttributes). Default namespaces must be given an explicit
name.}

\item{...}{Additional arguments passed down to methods. (Not currently used).}
}
\description{
This turns an XML document (or node or nodeset) into the equivalent R
list.
}
\details{
\code{as.list} currently only handles the four most common types of
children that an element might have:

\itemize{
  \item Other elements, converted to lists.
  \item Attributes, stored as R attributes.
  \item Text, stored as a character vector.
}
}
\examples{
as.list(xml("<foo> a <b /><c><![CDATA[<d></d>]]></c></foo>"))
as.list(xml("<foo> <bar><baz /></bar> </foo>"))
as.list(xml("<foo id = 'a'></foo>"))
as.list(xml("<foo><bar id='a'/><bar id='b'/></foo>"))
}

