% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_xml.R
\name{as_xml}
\alias{as_xml}
\title{Coerce a R list to xml nodes.}
\usage{
as_xml(x, ...)
}
\arguments{
\item{x}{A document, node, or node set.}

\item{...}{Needed for compatibility with generic. Unused.}
}
\description{
This turns an R list into the equivalent XML document. Not all R lists will
produce valid XML, in particular there can only be one root node and all
child nodes need to be named (or empty) lists.
}
\examples{
as_xml(list(x = list()))

# Nesting multiple nodes
as_xml(list(foo = list(bar = list(baz = list()))))

# attributes are stored as R attributes
as_xml(list(foo = structure(list(), id = "a")))
as_xml(list(foo = list(
     bar = structure(list(), id = "a"),
     bar = structure(list(), id = "b"))))
}
