% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_list.R
\name{deduplicate}
\alias{deduplicate}
\title{Deduplicate named elements in a list}
\usage{
deduplicate(lst)
}
\arguments{
\item{lst}{A list potentially containing duplicate named elements}
}
\value{
A list with duplicate elements consolidated
}
\description{
Deduplicate named elements in a list
}
