% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{xml_document-class}
\alias{xml_document-class}
\alias{xml_missing-class}
\alias{xml_node-class}
\alias{xml_nodeset-class}
\title{Register S4 classes}
\description{
Classes are exported so they can be re-used within S4 classes, see \code{\link[methods:setOldClass]{methods::setOldClass()}}.
\itemize{
\item \code{xml_document}: a complete document.
\item \code{xml_nodeset}: a \emph{set} of nodes within a document.
\item \code{xml_missing}: a missing object, e.g. for an empty result set.
\item \code{xml_node}: a single node in a document.
}
}
\keyword{internal}
