% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_url.R
\name{url_parse}
\alias{url_parse}
\title{Parse a url into its component pieces.}
\usage{
url_parse(x)
}
\arguments{
\item{x}{A character vector of urls.}
}
\value{
A dataframe with one row for each element of \code{x} and
columns: scheme, server, port, user, path, query, fragment.
}
\description{
Parse a url into its component pieces.
}
\examples{
url_parse("http://had.co.nz/")
url_parse("http://had.co.nz:1234/")
url_parse("http://had.co.nz:1234/?a=1&b=2")
url_parse("http://had.co.nz:1234/?a=1&b=2#def")
}
