% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xml_parse.R
\name{xml}
\alias{xml}
\title{Create XML from a string.}
\usage{
xml(x, base_url = NULL)
}
\arguments{
\item{x}{an xml string}

\item{base_url}{A base url to use for relative links.}
}
\description{
Create XML from a string.
}
\examples{
xml("<foo> 123 </foo>")
xml("<foo> \\u00e5 </foo>")
}

