% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml2-package.R
\docType{package}
\name{xml2-package}
\alias{xml2}
\alias{xml2-package}
\title{xml2: Parse XML}
\description{
Bindings to 'libxml2' for working with XML data using a simple, consistent interface based on 'XPath' expressions. Also supports XML schema validation; for 'XSLT' transformations see the 'xslt' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://xml2.r-lib.org}
  \item \url{https://r-lib.r-universe.dev/xml2}
  \item Report bugs at \url{https://github.com/r-lib/xml2/issues}
}

}
\author{
\strong{Maintainer}: Jeroen Ooms \email{jeroenooms@gmail.com}

Authors:
\itemize{
  \item Hadley Wickham
  \item Jim Hester
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item R Foundation (Copy of R-project homepage cached as example) [contributor]
}

}
\keyword{internal}
