% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xml2_example}
\alias{xml2_example}
\title{Get path to a xml2 example}
\usage{
xml2_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
xml2 comes bundled with a number of sample files in its \sQuote{inst/extdata}
directory. This function makes them easy to access.
}
