% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{xml_add_child}
\alias{xml_add_child}
\title{Add a child node}
\usage{
xml_add_child(x, value, copy = TRUE)
}
\arguments{
\item{x}{a document, node or nodeset.}

\item{value}{node or nodeset.}

\item{copy}{whether to copy the \code{value} before replacing. If this is \code{FALSE}
then the node will be moved from it's current location.}
}
\value{
the added element
}
\description{
Add a child node
}

