% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{xml_add_sibling}
\alias{xml_add_sibling}
\title{Append a sibling onto another node}
\usage{
xml_add_sibling(x, value, copy = TRUE)
}
\arguments{
\item{x}{a document, node or nodeset.}

\item{value}{node or nodeset to replace with.}

\item{copy}{whether to copy the \code{value} before replacing. If this is \code{FALSE}
then the node will be moved from it's current location.}
}
\value{
the added element
}
\description{
Append a sibling onto another node
}

