% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{xml_comment}
\alias{xml_comment}
\title{Construct a comment node}
\usage{
xml_comment(content)
}
\arguments{
\item{content}{The comment content}
}
\description{
Construct a comment node
}
\examples{
x <- xml_new_document()
r <- xml_add_child(x, "root")
xml_add_child(r, xml_comment("Hello!"))
as.character(x)
}
