% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_name.R
\name{xml_name<-}
\alias{xml_name<-}
\title{Modify the (tag) name of an element}
\usage{
xml_name(x, ns = character()) <- value
}
\arguments{
\item{x}{A document, node, or node set.}

\item{ns}{ignored for assignment}

\item{value}{a character vector with replacement name.}
}
\description{
Modify the (tag) name of an element
}
