% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xml_name.R
\name{xml_name}
\alias{xml_name}
\title{The (tag) name of an xml element.}
\usage{
xml_name(x, ...)
}
\arguments{
\item{x}{A document, node, or node set.}

\item{...}{Additional arguments passed down to methods. (Not currently used).}
}
\value{
A character vector.
}
\description{
The (tag) name of an xml element.
}
\examples{
x <- xml_parse("<bar>123</bar>")
xml_name(x)

y <- xml_parse("<bar><baz>1</baz>abc<foo /></bar>")
z <- xml_children(y)
xml_name(xml_children(y))
}

