% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{xml_new_namespace}
\alias{xml_new_namespace}
\title{Create a new namespace and assign it to a node}
\usage{
xml_new_namespace(x, namespaces)
}
\arguments{
\item{x}{a node}

\item{namespaces}{A named character vector of prefixes and URIs for the
namespaces. If one element is unnamed it is used as the default namespace.}
}
\value{
the node (invisibly)
}
\description{
Create a new namespace and assign it to a node
}
\examples{
d <- xml_new_document(xml_new_node("sld", version = "1.1.0", xml_new_node("layer", xml_new_node("Name"))))
sld <- xml_find_one(d, "//sld")
name <- xml_find_one(d, "//Name")
xml_new_namespace(sld, c("http://www.o.net/sld", ogc = "http://www.o.net/ogc", se = "http://www.o.net/se"))
xml_set_namespace(name, "se")
}

