% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{xml_new_node}
\alias{xml_new_node}
\title{Create a new node}
\usage{
xml_new_node(.name, ...)
}
\arguments{
\item{.name}{name of the node.}

\item{...}{Either named attributes or child nodes to add to the new node.}
}
\value{
A \code{xml_node} object.
}
\description{
Create a new node
}

