% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xml_push.R
\name{xml_parse_con}
\alias{xml_parse_con}
\title{Stream Parse XML}
\usage{
xml_parse_con(con, n = 1000, verbose = TRUE)
}
\arguments{
\item{con}{A connection object}

\item{n}{number of bytes to read per iteration}

\item{verbose}{output some info to the console}
}
\description{
Stream parse XML from a connection interface.
}
\examples{
con <- url("http://www.xmlfiles.com/examples/cd_catalog.xml")
doc <- xml_parse_con(con)

#https
library(curl)
}

