% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{xml_replace<-}
\alias{xml_replace<-}
\title{Replace a node with another node}
\usage{
xml_replace(x, copy = TRUE) <- value
}
\arguments{
\item{x}{a document, node or nodeset.}

\item{copy}{whether to copy the \code{value} before replacing. If this is \code{FALSE}
then the node will be moved from it's current location.}

\item{value}{node or nodeset to replace with.}
}
\description{
Replace a node with another node
}

