% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xml_save.R
\name{xml_save}
\alias{xml_save}
\title{Save xml nodes to a character vector.}
\usage{
xml_save(x)
}
\arguments{
\item{x}{A document, node, or node set.}
}
\value{
A character vector.
}
\description{
Save xml nodes to a character vector.
}
\examples{
x <- xml("<bar><baz>1</baz>abc<foo /></bar>")
xml_save(x)

y <- xml("<bar><baz>1</baz>abc<baz><foo>2</foo></baz></bar>")
ns <- xml_find(y, "//baz")
xml_save(ns)
}

