% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xml_search.R
\name{xml_search}
\alias{xml_search}
\title{Search for matching xpath expressions.}
\usage{
xml_search(x, xpath, ...)
}
\arguments{
\item{x}{A document, node, or node set.}

\item{xpath}{A string containing a xpath expression.}

\item{...}{Additional arguments passed down to methods. (Not currently used).}
}
\description{
Search for matching xpath expressions.
}
\examples{
x <- xml("<foo><bar><baz/></bar><baz/></foo>")
xml_search(x, ".//baz")
xml_path(xml_search(x, ".//baz"))

# Note the difference between .// and //
# //  finds anywhere in the document (ignoring the current node)
# .// finds anywhere beneath the current node
(bar <- xml_search(x, ".//bar"))
xml_search(bar, ".//baz")
xml_search(bar, "//baz")
}

