% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_modify.R
\name{xml_set_namespace}
\alias{xml_set_namespace}
\title{Set the node's namespace}
\usage{
xml_set_namespace(.x, prefix = "", uri = "")
}
\arguments{
\item{.x}{a node}

\item{prefix}{The namespace prefix to use}

\item{uri}{The namespace URI to use}
}
\value{
the node (invisibly)
}
\description{
The namespace to be set must be already defined in one of the node's
ancestors.
}
