% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xml_text.R
\name{xml_text}
\alias{xml_text}
\title{Extract the text}
\usage{
xml_text(x, ...)
}
\arguments{
\item{x}{A document, node, or node set.}

\item{...}{Additional arguments passed down to methods. (Not currently used).}
}
\value{
A character vector, the same length as x.
}
\description{
Extract the text
}
\examples{
x <- xml_parse("<p>This is some text. This is <b>bold!</b></p>")
xml_text(x)
xml_text(xml_children(x))
}

