% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_url.R
\name{xml_url}
\alias{xml_url}
\title{The URL of an XML document}
\usage{
xml_url(x)
}
\arguments{
\item{x}{A node or document.}
}
\value{
A character vector of length 1. Returns \code{NA} if the name is
not set.
}
\description{
This is useful for interpreting relative urls with \code{\link[=url_relative]{url_relative()}}.
}
\examples{
catalog <- read_xml(xml2_example("cd_catalog.xml"))
xml_url(catalog)

x <- read_xml("<foo/>")
xml_url(x)
}
