% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_schema.R
\name{xml_validate}
\alias{xml_validate}
\title{Validate XML schema}
\usage{
xml_validate(x, schema)
}
\arguments{
\item{x}{A document, node, or node set.}

\item{schema}{an XML document containing the schema}
}
\value{
TRUE or FALSE
}
\description{
Validate an XML document against an XML 1.0 schema.
}
\examples{
# Example from https://msdn.microsoft.com/en-us/library/ms256129(v=vs.110).aspx
doc <- read_xml(system.file("extdata/order-doc.xml", package = "xml2"))
schema <- read_xml(system.file("extdata/order-schema.xml", package = "xml2"))
xml_validate(doc, schema)
}
