% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_xslt.R
\name{xml_xslt}
\alias{xml_xslt}
\title{Apply XSLT Stylesheet}
\usage{
xml_xslt(x, stylesheet)
}
\arguments{
\item{x}{A document, node, or node set.}

\item{stylesheet}{an xml document containing an XSL stylesheet}
}
\description{
Transform an XML document using an XSL stylesheet.
}

