% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xml_parse.R
\name{read_xml}
\alias{read_html}
\alias{read_xml}
\alias{read_xml.character}
\alias{read_xml.connection}
\alias{read_xml.raw}
\title{Read HTML or XML from a stream or file.}
\usage{
read_xml(file, encoding = "", ..., as_html = FALSE)

read_html(file, encoding = "", ...)

\method{read_xml}{character}(file, encoding = "", ..., as_html = FALSE)

\method{read_xml}{raw}(file, encoding = "", xml_url = "", ...,
  as_html = FALSE)

\method{read_xml}{connection}(file, encoding = "", n = 64 * 1024,
  verbose = FALSE, ..., xml_url = "", as_html = FALSE)
}
\arguments{
\item{file}{A file, url, a connection, or a raw vector. Character vectors are
passed on to libxml, which supports both local paths and urls.
Connections are read in a chunk at a time; this is slower, but allows
you to use any R connection.}

\item{encoding}{Specify a default encoding for the document. Unless
otherwise specified XML documents are assumed to be in UTF-8 or
UTF-16. If the document is not UTF-8/16, and lacks an explicit
encoding directive, this allows you to supply a default.}

\item{...}{Additional arguments passed on to methods.}

\item{as_html}{Optionally parse an xml file as if it's html.}

\item{xml_url}{When loading from a connection or a raw vector, this allows
you to specify a base url. Base urls are used to turn relative urls into
absolute.}

\item{n}{If \code{file} is a connection, the number of bytes to read per
iteration. Defaults to 16kb.}

\item{verbose}{When reading from a slow connection, this prints some
output on every iteration so you know its working.}
}
\description{
Read HTML or XML from a stream or file.
}
\examples{
# You can read from a url:
catalog <- read_xml("http://www.xmlfiles.com/examples/cd_catalog.xml")
me <- read_html("http://had.co.nz")

# Or read from a raw vector
raw <- charToRaw("<x><y/></x>")
read_xml(raw)

# Read an HTML file
}
\seealso{
\code{\link{xml}()} to create an xml object from a string.
}

