% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_url.R
\name{url_escape}
\alias{url_escape}
\alias{url_unescape}
\title{Escape and unescape urls.}
\usage{
url_escape(x, reserved = "")

url_unescape(x)
}
\arguments{
\item{x}{A character vector of urls.}

\item{reserved}{A string containing additional characters to avoid escaping.}
}
\description{
Escape and unescape urls.
}
\examples{
url_escape("a b c")
url_escape("a b c", "")

url_unescape("a\%20b\%2fc")
url_unescape("\%C2\%B5")
}
