% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xml_write.R
\name{write_xml}
\alias{write_xml}
\title{Write XML to disk.}
\usage{
write_xml(x, file, ...)
}
\arguments{
\item{x}{A document or node to write to disk. It's not possible to
save nodesets containing more than one nodes.}

\item{file}{Path to file to write.}

\item{...}{Other arguments used by methods.}
}
\description{
This writes out both XML and normalised HTML.
}
\examples{
h <- html("<p>Hi!</p>")

tmp <- tempfile(fileext = ".xml")
write_xml(h, tmp)
read_xml(tmp)
}

