% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_write.R
\name{write_xml}
\alias{write_xml}
\alias{write_html}
\title{Write XML to disk.}
\usage{
write_xml(x, file, ...)

write_html(x, file, ...)
}
\arguments{
\item{x}{A document or node to write to disk. It's not possible to
save nodesets containing more than one node.}

\item{file}{Path to file to write.}

\item{...}{additional arguments passed to methods.}
}
\description{
This writes out both XML and normalised HTML.
}
\examples{
h <- read_html("<p>Hi!</p>")

tmp <- tempfile(fileext = ".xml")
write_xml(h, tmp)
readLines(tmp)

write_html(h, tmp)
readLines(tmp)
}

