% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xml_parse.R
\name{xml}
\alias{xml}
\title{Parse XML string}
\usage{
xml(x, base_url = NULL, encoding = NULL, options = NULL)
}
\arguments{
\item{x}{an xml string}

\item{base_url}{A base url to use for relative links.}

\item{encoding}{A default encoding to use.}

\item{options}{Additional options passed on to \code{xmlRead*}.
For expert use only.}
}
\description{
Parse XML string
}
\examples{
xml("<foo> 123 </foo>")
}

