% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xml_parse.R
\name{xml}
\alias{html}
\alias{xml}
\title{Create HTML or XML from a string.}
\usage{
xml(x, base_url = NULL)

html(x, base_url = NULL)
}
\arguments{
\item{x}{A string containing XML or HTML.}

\item{base_url}{A base url to use for relative links.}
}
\value{
Both functions return a xml_doc
}
\description{
Create HTML or XML from a string.
}
\examples{
xml("<foo> 123 </foo>")
xml("<foo> \\u00e5 </foo>")

# The html parser converts potential malformed HTML into valid XML
html("<html> 123 </html>")
}

