% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml2-package.R
\docType{package}
\name{xml2-package}
\alias{xml2}
\alias{xml2-package}
\title{xml2: Parse XML}
\description{
Work with XML files using a simple, consistent interface. Built on top of the 'libxml2' C library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://xml2.r-lib.org/}
  \item \url{https://github.com/r-lib/xml2}
  \item Report bugs at \url{https://github.com/r-lib/xml2/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co}

Authors:
\itemize{
  \item Jim Hester
  \item Jeroen Ooms
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item R Foundation (Copy of R-project homepage cached as example) [contributor]
}

}
\keyword{internal}
