% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xml_attr.R
\name{xml_attr}
\alias{xml_attr}
\alias{xml_attr_exists}
\alias{xml_attrs}
\title{Retrieve an attribute.}
\usage{
xml_attr(x, attr)

xml_attr_exists(x, attr)

xml_attrs(x, ns = character())
}
\arguments{
\item{x}{A document, node, or node set.}

\item{attr}{Name of attribute to extract.}

\item{ns}{Optionally, a named vector giving prefix-url pairs, as produced
by \code{\link{xml_ns}}. If provided, all names will be explicitly
qualified with the ns prefix, i.e. if the element \code{bar} is defined
in namespace \code{foo}, it will be called \code{foo:bar}. (And
similarly for atttributes). Default namespaces must be given an explicit
name.}
}
\value{
For \code{xml_attr}, a character vector. If an attribute is not
 presented, its value will be missing. For \code{xml_attr_exists},
 a logical vector. For \code{xml_attr} a list of named character vectors.
 If any attrbutes have an associated namespace, the vector will have
 a \code{ns} attribute.
}
\description{
Use \code{xml_attrs} to retrieve all attributes values as a character vector.
Use \code{xml_attr} to retrieve the value of single attribute. If the
attribute doesn't exist, it will be returned as an \code{NA}.
Use \code{xml_attr_exists} to test if an attribute is present.
}
\examples{
x <- xml("<root id='1'><child id ='a' /><child id='b' d='b'/></root>")
xml_attr(x, "id")
xml_attrs(x)

xml_attr(xml_children(x), "id")
xml_attr_exists(xml_children(x), "id")
xml_attrs(xml_children(x))

# Missing attributes give missing values
xml_attr(xml_children(x), "d")
xml_attr_exists(xml_children(x), "d")
}

