% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xml_attr.R
\name{xml_attr}
\alias{xml_attr}
\title{Retrieve an attribute.}
\usage{
xml_attr(x, attr, ...)
}
\arguments{
\item{x}{A document, node, or node set.}

\item{attr}{Name of attribute to extract.}

\item{...}{Additional arguments passed down to methods. (Not currently used).}
}
\value{
A character vector. If an attribute is not presented, its
 value will be missing.
}
\description{
Retrieve an attribute.
}
\examples{
x <- xml_parse("<root id='1'><child id ='a' /><child id='b' d='b'/></root>")
xml_attr(x, "id")
xml_attr(xml_children(x), "id")
xml_attr(xml_children(x), "d")
}

