% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{xml_cdata}
\alias{xml_cdata}
\title{Construct a cdata node}
\usage{
xml_cdata(content)
}
\arguments{
\item{content}{The CDATA content, does not include \verb{<![CDATA[}}
}
\description{
Construct a cdata node
}
\examples{
x <- xml_new_root("root")
xml_add_child(x, xml_cdata("<d/>"))
as.character(x)
}
