% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xml_children.R
\name{xml_children}
\alias{xml_children}
\title{Extract children.}
\usage{
xml_children(x, ...)
}
\arguments{
\item{x}{A document, node or nodeset.}

\item{...}{Additional arguments passed down to methods. (Not currently used).}
}
\value{
A nodeset
}
\description{
Extract children.
}
\examples{
x <- xml("<foo> <bar><boo /></bar> 123 <baz/> </foo>")
(kids <- xml_children(x))
(grand_kids <- xml_children(kids))
}

