% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xml_find.R
\name{xml_find}
\alias{xml_find}
\title{Search for matching xpath expressions.}
\usage{
xml_find(x, xpath, ns = character(), ...)
}
\arguments{
\item{x}{A document, node, or node set.}

\item{xpath}{A string containing a xpath expression.}

\item{ns}{Optionally, a named vector giving prefix-url pairs, as produced
by \code{\link{xml_ns}}. If provided, all names will be explicitly
qualified with the ns prefix, i.e. if the element \code{bar} is defined
in namespace \code{foo}, it will be called \code{foo:bar}. (And
similarly for atttributes). Default namespaces must be given an explicit
name.}

\item{...}{Additional arguments passed down to methods. (Not currently used).}
}
\description{
Search for matching xpath expressions.
}
\examples{
x <- xml("<foo><bar><baz/></bar><baz/></foo>")
xml_find(x, ".//baz")
xml_path(xml_find(x, ".//baz"))

# Note the difference between .// and //
# //  finds anywhere in the document (ignoring the current node)
# .// finds anywhere beneath the current node
(bar <- xml_find(x, ".//bar"))
xml_find(bar, ".//baz")
xml_find(bar, "//baz")
}

