% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{xml_new_document}
\alias{xml_new_document}
\title{Create a new document}
\usage{
xml_new_document(node, version = "1.0")
}
\arguments{
\item{node}{The root node of the document.}

\item{version}{The version number of the document.}
}
\description{
Create a new document
}

