% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xml_path.R
\name{xml_path}
\alias{xml_path}
\title{Retrieve the xpath to a node}
\usage{
xml_path(x, ...)
}
\arguments{
\item{x}{A document, node, or node set.}

\item{...}{Additional arguments passed down to methods. (Not currently used).}
}
\value{
A character vector.
}
\description{
This is useful when you want to figure out where nodes matching an
xpath expression live in a document.
}
\examples{
x <- xml("<foo><bar><baz /></bar><baz /></foo>")
xml_path(xml_search(x, ".//baz"))
}

