% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_type.R
\name{xml_type}
\alias{xml_type}
\title{Determine the type of a node.}
\usage{
xml_type(x)
}
\arguments{
\item{x}{A document, node, or node set.}
}
\description{
Determine the type of a node.
}
\examples{
x <- read_xml("<foo> a <b /> <![CDATA[ blah]]></foo>")
xml_type(x)
xml_type(xml_contents(x))
}
