% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_xslt.R
\name{xml_xslt}
\alias{xml_xslt}
\title{XSLT 1.0 Transformations}
\usage{
xml_xslt(x, stylesheet)
}
\arguments{
\item{x}{A document, node, or node set.}

\item{stylesheet}{an xml document containing an XSL stylesheet}
}
\description{
Transform an XML document by applying an XSL stylesheet.
}
\details{
This implementation supports XSLT 1.0 features plus most of the EXSLT set of
processor-portable extensions functions. Unfortunately XSLT 2.0 or 3.0 features
are only available in proprietary libraries and currently unsupported. However
XSLT 2.0 is not widely adopted anyway because it is also unavailable in most
browsers.
}

